flask-api/requirements.txt
Flask==2.2.5
Flask_Cors==5.0.0
h5py==3.12.1
joblib==1.4.2
numpy==2.2.4
pandas==2.2.3
python-dotenv==1.1.0
pytz==2024.1
Requests==2.32.3
scikit_learn==1.6.1
tensorflow==2.18.0
torch==2.3.1+cpu
